requirements.txt
absl-py==2.3.1
accelerate==1.13.0
aiofiles==23.2.1
altair==5.5.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
archspec @ file:///croot/archspec_1709217642129/work
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
causal_conv1d==1.6.0
certifi @ file:///home/task_176765899978398/croot/certifi_1767659142032/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.3.1
comm==0.2.3
conda @ file:///croot/conda_1736277811838/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1714483155348/work
conda_package_streaming @ file:///work/perseverance-python-buildout/croot/conda-package-streaming_1698847176583/work
contourpy==1.3.3
cryptography @ file:///croot/cryptography_1714660666131/work
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.27.2
distro @ file:///croot/distro_1714488253808/work
einops==0.8.2
executing==2.2.0
fastapi==0.135.1
fastjsonschema==2.21.1
ffmpy==1.0.0
filelock==3.18.0
fonttools==4.59.0
fqdn==1.5.1
frozendict @ file:///home/task_176175050888873/conda-bld/frozendict_1761750665347/work
fsspec==2025.7.0
fvcore==0.1.5.post20221221
gradio==3.41.2
gradio_client==0.5.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.1
idna @ file:///croot/idna_1714398848350/work
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab-language-pack-zh-CN==4.4.post0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.2
kornia_rs==0.1.10
lark==1.2.2
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
lpips==0.1.4
# Editable install with no version control (mamba_ssm==1.2.0.post1)
-e /root/autodl-tmp/ShadowChallenge/networks/mamba-ssm-main
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1714510563922/work
mistune==3.1.3
mkl-service==2.5.2
mpmath==1.3.0
narwhals==2.18.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
notebook_shim==0.2.4
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.92
orjson==3.11.7
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.18.1
pexpect==4.9.0
pillow==11.3.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
portalocker==3.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptflops==0.7.5
ptyprocess==0.7.0
pure_eval==0.2.3
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2026.1.15
requests @ file:///croot/requests_1707355572290/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.26.0
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.0
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==69.5.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.52.1
supervisor==4.2.5
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
terminado==0.18.1
thop==0.1.1.post2209072238
tifffile==2026.1.28
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.15.2
torch @ https://download.pytorch.org/whl/cu128/torch-2.8.0%2Bcu128-cp312-cp312-manylinux_2_28_x86_64.whl#sha256=4354fc05bb79b208d6995a04ca1ceef6a9547b1c4334435574353d381c55087c
torchvision @ https://download.pytorch.org/whl/cu128/torchvision-0.23.0%2Bcu128-cp312-cp312-manylinux_2_28_x86_64.whl#sha256=9cb3c13997afcb44057ca10d943c6c4cba3068afde0f370965abce9c89fcffa9
tornado==6.5.1
tqdm @ file:///croot/tqdm_1714567712644/work
traitlets==5.14.3
transformers==4.36.2
triton==3.4.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typer==0.24.1
typer-slim==0.21.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.3
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1707770551213/work
uvicorn==0.41.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.14
yacs==0.1.8
zipp==3.23.0
zstandard @ file:///croot/zstandard_1714677652653/work
